eureka
huh , I made a mistake and my program did not work as i planned.
Namely, I declared only one one variable for choice 2, choosing the
area or volume of a figure. I should have made a new one for every
choosing; one for the cube, one for the prism, one for the sphere…
I corrected this, but before I did the program asked me for a,b,c,r,s
for A CUBE! hahahh
now the mistake is removed, and the program works as i wanted it to
work :DD